Improves object creation count when error list entries changes.
Until now every entries changed event of the ErrorList build up a new list of DiagnosticInfos, which was build from ground up, which increases GC pressure by a lot.
Now we compare hashes in beforehand to only build new DiagnosticInfo objects, when really necessary.
We also could think of a way of pooling Diagnostic Infos, but i tend do make them immutable in the future, which would prevent pooling..

Also fixed a small bug in the placement calculation, which occurred when lines near the end of the document where deleted and the last diagnostics in the document went out of range.